+++ tachyon/math/finite_fields/fp3.h
@@ -27,7 +27,7 @@ class Fp3 final : public CubicExtensionField<Fp3<Config>> {
   static_assert(Config::kDegreeOverBaseField == 3);
   static_assert(BaseField::ExtensionDegree() == 1);
 
-  constexpr static uint64_t kDegreeOverBasePrimeField = 3;
+  constexpr static uint32_t kDegreeOverBasePrimeField = 3;
 
   static void Init() {
     Config::Init();
